Move rule generated-flow-types to own file
